Fixed some bugs
The five cards being shown as a preview are now guaranteed to be all
different from one another. There will never be a repeated card.

When the user clicks too many selectors to change the preview card at
once, it used to bug out and keep a flash card shown on screen. Now, it
can handle multiple clicks and even saves the user's clicks so that it
can go the last page the user clicked. If a user clicks pages 1, 2, then
3, the flash cards will go to page 1 first (since that was the first
page the user clicked), then go straight to page 3 since the user ended
on page 3.

Added google analytics script

The entire script is now in an anonymous function to prevent global
variables