Created the study page for acadec 2017-2018
This is where the user can take practice tests to study for acadec using the data in 'questions.json'. One will probably be added for 2019 when that year rolls around, but for now just 2017-2018.
This webpage was thrown together in 2 days, so it's not compatible with really old browsers unless I fix it. Just for now though, it works on basically any average person's browser.