Added lang, <!DOCTYPE>, and auto-hypens
I forgot to add a lang and <!DOCTYPE html> declaration on the last commit, so I added them here. I also added hypens: auto; to some elements where there's a lot of text just to make it better when lines wrap.
Also The "Social Science" button shouldn't be able to be the only button to wrap (since it's the only one with a space), so it's space has been changed to a non-breaking space.