Added a kinda latex-ish parser
Since the math section is going to need some symbols that aren't in the normal font, like "lim" with text under it, fractions, and combination syntax (2 numbers inside of parentheses, like a fraction with no bar), I've added a parser that will convert certain syntax into symbols. So far, only \lim and \frac are implemented, but I'll add more later.
As an example:
"a fraction -- $\frac{03|100}{08|<i>x</i>}$" will render the string "a fraction -- " + a fraction with 100 in the numerator and an italicized x in the denominator.
"a limit -- $\lim{08|<i>x</i>}{01|4}$" will render the string "a limit -- " + the word "lim" with "x➙4" underneath it (x is italicized again) (that's the closest arrow I could find that doesn't look out of place with the Passion One font).
It's different from latex, but still kind of uses the same basic syntax.
Also, arguments can contain nested latex or html that will be parsed and displayed correctly.