Added support for \binom
For math problems pertaining to probability, "\binom" can be used to create a binomial (a pair of parentheses around two numbers stacked on top of each other) representing the function for combinations (nCr). Also, parentheses can be wrapped around a block of text and will match the height of the text using "\around". It's basically a stretched SVG with one <text> element.