Update 2018.html
The timer automatically pauses at 3:00 now for each question. It's to prevent leaving a question unanswered and then coming back to have the timer be saying like two hours because it was running the whole time. Three minutes is plenty to answer a question, so it shouldn't really affect any real times. It also means the maximum amount of time per test is two and half hours, and that's if someone waits three minutes or more for every single question.
The timer will also show hours now is someone manages to take that long. Under one hour, only the minutes and seconds will be shown. After an hour though, the minutes will be reset and an extra ":" will be added to the front to separate the new hour number.
The "End Test Now" button was changed to a "Quit Test" button just because it looks better. It still ends the test before reaching all 50 questions like normal.