Favicon and more mobile friendly
All the favicon stuff has been added in.
".../dice" should now be a bit more mobile-friendly:
When you save the page on an Apple device, it'll open in application
mode (no Safari bar blocking the page).
For some reason, the two boxes that separate the page would be offset
from each other even though they're supposed to be at the same height.
That's been fixed by making their parent a flex box.
The middle bar you drag to resize the two boxes didn't really work on
touch devices. It was too small and would cause the whole page to scroll
instead. It's size should now be proportional to the height of the
screen and it shouldn't scroll anymore while it's being dragged. It also
uses "vh" units now so that the bar is always at the same height as the
visible line separator.
iOS Safari didn't have a glyph for those little triangles on the saved
rolls menu. The triangle has been changed into another, similar-looking
character that does have a glyph on Safari. It looks a little bigger
than the old one, but a big glyph is better than no glyph.