Better stats integration
Now, instead of trying to get the min and max of a roll from
`dieroller.js', the stats of a roll are sent to stats.js to be figured
out. For most rolls, the calculation is almost instant anyway, so not
that big of an extra wait. The reason for that is that I falsely assumed
the minimum of a roll would always be if every dice rolled a 1. Consider
"-3d20". Previously, the maiinimum would have come out to -3, even
though the real minimum is -60.