More functions, new logo, and bug fixes.
- Some single-argument functions have been added, like abs() or int().
For average dice rolls, you probably won't need them, but they're there
just in case. Full list: abs, round, ceil, floor, sign, trunc (and their
aliases).
- The old logo kind of sucked. The new one isn't that much different,
but overall, it does look a little better. Instead of lines separating
the faces on the die, the faces are now slightly different shades to
give it a shadow-y effect and make the separation more clear.
- When a dice roll's stats get calculated, the cache is checked first so
that the stats for the same roll don't have to be calculated each time.
It makes getting stats a lot faster and less redundant. But for rolls
where the stats took too long, that result (that there are no results
because it's too big) wasn't being saved. So every time the stats were
being calculated, any rolls where the stats took too long would try
again to re-calculate them, even though obviously, it won't change: the
stats would still take too long to get because the roll hadn't changed.
Now, that's stored in the cache too so it doesn't try to get the stats
of a huge roll every time. It tries once and stops if it fails.
- A couple minor bugs were also fixed.