Updated homepage, deleted old stuff, and add GitHub Pages workflow
The old homepage from like 8th grade has been replaced with more recent stuff. A lot of the junk I made when I was just learning to code has been removed since I never use it. Instead of deploying from the master branch, the entire branch is built with jekyll, and the generated site files are cloned to the gh-pages branch, where GitHub now knows to look for the actual website.