Changed look of website
I completely changed the styles of the website to look better, especially the front page. There are also a couple other minor changes, but the main thing is that the main index.html page changed completely and the rest of the website will probably updated over time to match those styles.